requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-WTF>=1.0
Flask-Login>=0.5
Flask-Migrate>=3.0 # Para manejar cambios en la DB
Flask-Mail>=0.9    # Para enviar correos
Werkzeug>=2.0      # Para hashing de contraseñas
python-dotenv>=0.19 # Para variables de entorno locales
email_validator>=1.1 # Para validar emails en formularios
psycopg2-binary    # Driver de PostgreSQL para Render (importante)
gunicorn           # Servidor WSGI para producción